requirements-support.txt
# Support Chat Feature Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
